PASSES EVERYTHING IN LAB1

+++ Lab1/src/java/simpledb/HeapPage.java
@@ -86,7 +86,7 @@ public class HeapPage implements Page {
      * @return the number of bytes in the header of a page in a HeapFile with each tuple occupying tupleSize bytes
      */
     private int getHeaderSize() {        
-        return (int) Math.ceil(getNumTuples()/8);
+        return (int) Math.ceil((float)getNumTuples()/8);
                  
     }
